requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.0
cssselect2==0.4.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0
fonttools==4.28.4
future==0.18.2
gcloud==0.18.3
googleapis-common-protos==1.54.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
jwcrypto==1.0
lxml==4.7.1
oauth2client==4.1.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
pydyf==0.1.2
pyparsing==3.0.6
PyPDF2==1.26.0
pyphen==0.11.0
Pyrebase4==4.5.0
python-bidi==0.4.2
python-firebase==1.2
python-jwt==3.3.1
qrcode==7.3.1
razorpay==1.2.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
tzdata==2021.5
urllib3==1.26.7
webencodings==0.5.1
wrapt==1.13.3
zopfli==0.1.9
dash-bootstrap-components==0.10.1
